feat: Create flux store and actions
The UI of the control panel now updates dynamically when the username of
the font size fields are changed on the UI.

I kind of went off on a limb and implemented Actions in such a way that
everything didn't have to been the same file.  This is a much more
scalable solution than was was presented in the tutorial, however the
factories that I have created are rather verbose to do something as
simple as get the type.  There is probably a better way, but this is
fine for now.